// packages/fte.js-base/src/types/BlockContent.ts
import { BlockRunFunction } from './BlockRunFunction'
import { ContentFunction } from './ContentFunction'
import { DefaultFactoryOption } from './DefaultFactoryOption'
import { PartialFunction } from './PartialFunction'
import { SlotFunction } from './SlotFunction'
import { SlotsHash } from './SlotsHash'

export type BlockContent<OPTIONS extends DefaultFactoryOption = DefaultFactoryOption> = {
  partial: PartialFunction
  content: ContentFunction
  run: BlockRunFunction
  slots: SlotsHash
  slot: SlotFunction
  options: OPTIONS
}